smolagents-examples/basic/pyproject.toml
[project]
name = "basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "smolagents>=1.14.0",
    "wikipedia-api>=0.8.1",
]
